logica overzicht + bug fix oefeningen
Na keuze labo worden de oefeningen getoond, controle werkt bij beide type oefingen correct
