+++ src/pages/home/home.html
@@ -7,11 +7,5 @@
 </ion-header>
 
 <ion-content padding>
-  <button *ngFor="let labo of labos" ion-button (click)="controle(labo)" >
-         {{labo}}
-   </button>
-
-   <ion-list >
-       <ion-item *ngFor="let stap of stappen">{{ stap }}</ion-item>
-   </ion-list>
+  
 </ion-content>
