+++ src/pages/labotafel/labotafel.html
@@ -16,7 +16,7 @@
 
 
 <ion-content padding class="background">
-
+<ion-title>{{vraag}}</ion-title>
 
   <div *ngFor="let item of materiaalGeselecteerd">
     <ion-item>{{item}}</ion-item>
