+++ src/pages/materiaal-selecteren/materiaal-selecteren.html
@@ -6,7 +6,7 @@
 -->
 <ion-header>
 
-  <ion-navbar>
+  <ion-navbar hideBackButton >
     <ion-title>MateriaalSelecteren</ion-title>
   </ion-navbar>
 
@@ -15,18 +15,19 @@
 
 <ion-content padding>
 
-  <button ion-button (click)="pushPage()">
-         Labotafel
-  </button>
+<ion-title>{{vraag}}</ion-title>
 
   <ion-list >
 
   <ion-item *ngFor="let item of materiaal">
-    <ion-label>{{item.name}}</ion-label>
-    <ion-checkbox [(ngModel)]="checkItems[item.name]"  value="item.name"  (ionChange)="print()"></ion-checkbox>
+    <ion-label>{{item}}</ion-label>
+    <ion-checkbox [(ngModel)]="checkItems[item]"  value="item"  (ionChange)="print()"></ion-checkbox>
   </ion-item>
 
 <!-- ionChange: Call function when state changes [(ngModel)]="checkItems[item.naam]" -->
+  <button ion-button (click)="pushPage()">
+       Labotafel
+     </button>
 
   </ion-list>
 </ion-content>
