+++ src/pages/overzicht/overzicht.html
@@ -1,0 +1,24 @@
+<!--
+  Generated template for the OverzichtPage page.
+
+  See http://ionicframework.com/docs/components/#navigation for more info on
+  Ionic pages and navigation.
+-->
+<ion-header>
+
+  <ion-navbar>
+    <ion-title>overzicht</ion-title>
+  </ion-navbar>
+
+</ion-header>
+
+
+<ion-content padding>
+  <button *ngFor="let labo of labos" ion-button (click)="controle(labo)" >
+         {{labo}}
+   </button>
+
+   <ion-list >
+       <ion-item *ngFor="let stap of stappen">{{ stap }}</ion-item>
+   </ion-list>
+</ion-content>
