+++ src/pages/overzicht/overzicht.scss
@@ -1,0 +1,3 @@
+page-overzicht {
+
+}
